*Added CAN reading code to a new thread *Added CAN frame as a data member for message node
