--- Debug/BikeAVL.d
@@ -1,1 +1,0 @@
-BikeAVL.d: ../BikeAVL.c
