--- Debug/avl.d
@@ -1,27 +1,0 @@
-avl.d: ../avl.c ../all.h ../include/socketcan/can.h ../lib.h ../parser.h \
- ../avl.h ../MessageAVL.h ../LinkedList.h ../signal.h ../SignalAVL.h \
- ../datalogger.h ../translator.h ../syslog.h
-
-../all.h:
-
-../include/socketcan/can.h:
-
-../lib.h:
-
-../parser.h:
-
-../avl.h:
-
-../MessageAVL.h:
-
-../LinkedList.h:
-
-../signal.h:
-
-../SignalAVL.h:
-
-../datalogger.h:
-
-../translator.h:
-
-../syslog.h:
