--- Debug/sources.mk
@@ -1,17 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-O_SRCS := 
-C_SRCS := 
-S_UPPER_SRCS := 
-OBJ_SRCS := 
-ASM_SRCS := 
-OBJS := 
-C_DEPS := 
-EXECUTABLES := 
-
-# Every subdirectory with source files must be described here
-SUBDIRS := \
-. \
-
