+++ headers/all.h
@@ -42,7 +42,7 @@
 #include <linux/can/error.h>
 
 
-#include "include/socketcan/can.h"
+//#include "socketcan/can.h"
 #include "lib.h"
 #include "parser.h"
 #include "avl.h"
@@ -54,6 +54,6 @@
 #include "syslog.h"
 #include "signal.h"
 
-
+#include "can_interceptor.h"
 
 #endif /* ALL_H_ */
