+++ headers/can_interceptor.h
@@ -1,0 +1,13 @@
+/*
+ * can_interceptor.h
+ *
+ *  Created on: Jan 7, 2015
+ *      Author: cancorder
+ */
+
+#ifndef CAN_INTERCEPTOR_H_
+#define CAN_INTERCEPTOR_H_
+
+int * can_interceptor_thread();
+
+#endif /* CAN_INTERCEPTOR_H_ */
